Draws the Circle Now
it draws the circle, but only the red slider works. The green one does absolutely  nothing, and it's due to the hex code not putting anything in about green. Lines 80-81 add a string of 0's to complete the hex code which is why the green is uncontrolled, but when those 0's arent there, it doesn't draw the circle at all.

-------------Also -------------

For some reason, the text box on the right of the GUI no longer displays the hex code, but is blank even though I didn't change any of the code of the text box.